backend/requirements.txt
# Django Core
Django==4.2.16
djangorestframework==3.14.0

# Database Drivers
psycopg2-binary==2.9.9

# Environment Variables
python-decouple==3.8

# JWT Authentication
djangorestframework-simplejwt==5.3.1

# CORS
django-cors-headers==4.3.1

# Filtering
django-filter==23.5

# API Documentation
drf-spectacular==0.27.0

# Redis Cache
django-redis==5.4.0

# Development Tools
django-debug-toolbar==4.2.0
django-extensions==3.2.3

# Production Server
gunicorn==21.2.0

# Mercado Pago (para futura integración)
mercadopago==2.0.0

